requirements.txt
bcrypt
libsql-client
python-dotenv
